refactor: move test folder outside of src
