+++ tsconfig.json
@@ -13,6 +13,6 @@
       "forceConsistentCasingInFileNames": true,
       "moduleResolution": "node"
     },
-    "include": ["src", "*.config.js"],
+    "include": ["src", "*.config.js", "test"],
     "exclude": ["node_modules"]
   }
